refactor: consolidate meals feature into nutrition module
This commit merges the standalone `meals` feature module into the `nutrition` module to unify health-related nutritional logic. It reorganizes the codebase into ritual-specific and shared components.

- **Module Consolidation**:
    - Deleted the `:features:health:meals` module and moved its source code into `:features:health:nutrition`.
    - Updated `settings.gradle.kts` and project build files to reflect the removal of the `meals` module.
- **Package Restructuring**:
    - Relocated meal data models and `MealsRepository` to `com.zoewave.probase.features.health.nutrition.data`.
    - Restructured UI logic into `ui.ritual` (for specific ritual flows) and `ui.shared` (for reusable components like `BioMealCard`, `EditMealScreen`, and `MealPreparationScreen`).
- **Dependency Management**:
    - Transferred CameraX, Coil, and Accompanist permission dependencies to `:features:health:nutrition` to support meal capture and image rendering features.
- **Application Integration**:
    - Updated `FeatureInventoryEntryProvider` and `KoColorNavEntryProvider` to use the new nutrition package paths.
    - Refactored imports in the `routines` feature to reference shared nutrition components.
- **Cleanup**:
    - Removed redundant walkthrough documentation and build artifacts from the deleted `meals` module.
