+++ app/src/main/java/com/zoewave/probase/ui/components/FeatureInventoryEntryProvider.kt
@@ -13,7 +13,7 @@ import com.zoewave.probase.features.ble.ui.BluetoothLeRoute
 import com.zoewave.probase.features.calendar.ui.CalendarUiRoute
 import com.zoewave.probase.features.camera.ui.CameraUIRoute
 import com.zoewave.probase.features.health.core.ui.HealthRoute
-import com.zoewave.probase.features.health.meals.ui.MealsUiRoute
+import com.zoewave.probase.features.health.nutrition.ui.shared.MealsUiRoute
 import com.zoewave.probase.features.nav3.ui.inventory.FeatureInventory
 import com.zoewave.probase.features.nav3.ui.inventory.FeatureInventoryScreen
 import com.zoewave.probase.features.microphone.SpeechTestScreen
