+++ applications/kocolor/features/routines/build.gradle.kts
@@ -17,7 +17,7 @@ dependencies {
     implementation(project(":applications:kocolor:db"))
     implementation(project(":applications:kocolor:data"))
     implementation(project(":features:xr:glass"))
-    implementation(project(":features:health:meals"))
+    implementation(project(":features:health:nutrition"))
 
     implementation(libs.androidx.projected)
     implementation(libs.androidx.compose.material.icons.extended)
