+++ applications/kocolor/features/routines/src/main/java/com/zoewave/probase/kocolor/features/routines/ui/components/StepHeroPage.kt
@@ -55,7 +55,7 @@ import coil.compose.AsyncImage
 import com.zoewave.probase.core.model.ritual.CosmeticItem
 import com.zoewave.probase.core.model.ritual.RoutineStep
 import com.zoewave.probase.core.model.ritual.RoutineTime
-import com.zoewave.probase.features.health.meals.ui.components.BioMealSuggestionSection
+import com.zoewave.probase.features.health.nutrition.ui.shared.BioMealSuggestionSection
 import com.zoewave.probase.kocolor.features.routines.R
 import com.zoewave.probase.kocolor.features.routines.ui.RoutinesEvent
 import com.zoewave.probase.kocolor.model.KoColorRoute
