--- features/health/meals/build.gradle.kts
@@ -1,37 +1,0 @@
-plugins {
-    id("composetemplate.android.library")
-    id("composetemplate.android.library.compose")
-    id("composetemplate.android.hilt")
-    id("composetemplate.kotlin.serialization")
-}
-
-android {
-    namespace = "com.zoewave.probase.features.health.meals"
-}
-
-dependencies {
-    implementation(project(":core:ui"))
-    implementation(project(":core:model"))
-    implementation(project(":core:data"))
-    
-    implementation(libs.androidx.projected)
-    implementation(libs.androidx.navigation3.runtime)
-    implementation(libs.androidx.navigation3.ui)
-    implementation(libs.androidx.lifecycle.viewmodel.navigation3)
-    implementation(libs.androidx.hilt.lifecycle.viewmodel.compose)
-    implementation(libs.kotlinx.serialization.json)
-    implementation(libs.androidx.compose.material.icons.extended)
-
-    // CameraX
-    implementation(libs.androidx.camera.core)
-    implementation(libs.androidx.camera.camera2)
-    implementation(libs.androidx.camera.lifecycle)
-    implementation(libs.androidx.camera.view)
-    implementation(libs.androidx.camera.compose)
-    
-    // Coil for images
-    implementation(libs.coil.compose)
-
-    // Permission handling
-    implementation(libs.google.accompanist.permissions)
-}
