--- features/health/meals/meals_management_walkthrough.artifact.md
@@ -1,30 +1,0 @@
-# Walkthrough - Management Capabilities for Metabolic Meals
-
-I have successfully implemented a comprehensive management system for the **Metabolic Meals** feature, allowing you to edit, delete, and fully curate your captured nutritional protocols.
-
-## Key Accomplishments
-
-### 1. Full CRUD Operations
-- **Meal Editing**: Introduced a high-fidelity **`EditMealScreen`** that allows you to modify every detail of a captured meal, including:
-    - **Protocol Name & Focus**: Edit the scientific title and biological rationale.
-    - **Metabolic Phase**: Re-categorize meals between Morning (mTOR), MidDay (Microbiome), and Evening (Autophagy) windows.
-    - **Nutritional Matrix**: Update precise values for Calories, Protein, Carbs, and Fats.
-- **Secure Deletion**: Added a delete action with a stylized **confirmation dialog** to prevent accidental removal of your metabolic history.
-
-### 2. "Bio-Optimized" Editing Interface
-- **Phase-Adaptive Styling**: The edit form dynamically updates its accent colors (**Lime**, **Cyan**, **Pink**) based on the selected metabolic phase, ensuring visual consistency with the "Atelier" design language.
-- **Architectural Layout**: Used non-clipping shadows and deep slate backgrounds to maintain the high-end, immersive feel of the feature.
-
-### 3. Refined User Flow
-- **Interactive Detail View**: Upgraded the `MealDetailScreen` with intuitive Edit and Delete actions in the top app bar.
-- **Seamless State Sync**: The `MealsViewModel` now reactively handles updates and deletions, ensuring the list, detail, and edit views are always perfectly synchronized.
-
-## Technical Details
-- **Architecture**: Expanded the `MealsRepository` and `MealsViewModel` to support reactive state updates for updates and deletions.
-- **Build Status**: Verified with a successful clean build: `:features:health:meals:assembleDebug`.
-
----
-> [!SUCCESS]
-> Your Metabolic Meals are now fully manageable. Open any captured meal to refine its scientific data or remove it from your archive.
-
-**KoColor now provides a complete, high-precision lifecycle for your nutritional optimization protocols.**
