--- features/health/meals/meals_ritual_prep_walkthrough.artifact.md
@@ -1,33 +1,0 @@
-# Walkthrough - Meals Ritual Suggestion & Preparation Flow
-
-I have successfully integrated the **Meal Suggestion** cards into the **Meals Ritual** and implemented a high-fidelity **Preparation Flow** to guide users through their metabolic protocols.
-
-## Key Accomplishments
-
-### 1. Integrated Meal Suggestions
-- **Reusable Component**: Created the **`MealSuggestionSection`** within the `:features:health:meals` module. This component is now used directly within the Ritual Knowledge hub.
-- **Choose Your Protocol**: Users can now browse multiple metabolic meal options (e.g., mTOR Activation Breakfast, SCFA Lunch) directly from their ritual stages.
-- **Persistent Selection**: Tapping a meal card now links that specific protocol to the current ritual stage, providing a personalized Bio-Sync experience.
-
-### 2. High-Fidelity Preparation Flow
-- **Interactive Step-by-Step**: Developed the **`MealPreparationScreen`**, a spectacular new interface that guides users through the cooking process.
-- **Visual Progress Timeline**: Features a vertical "Step Timeline" with checkmarks and active "In Progress" states, matching your spectacular design.
-- **Contextual Guidance**: Each step includes a "Previous" and "Next Step" navigation bar, ensuring users stay focused on the biochemical execution of their meal.
-
-### 3. "Start Cooking" Detail View
-- **Actionable Details**: Upgraded the `MealDetailScreen` with a prominent **"Start Cooking"** button, creating a seamless transition from scientific research to practical execution.
-- **Phase-Synced Styling**: Maintained the Bio-Optimized color system (Lime, Cyan, Pink) throughout the entire preparation journey.
-
-### 4. Deep-Link Navigation
-- **Unified Routing**: Updated the `MealsHub` route to support deep-linking. Clicking a meal suggestion in the ritual Hub now jumps you directly into the **Preparation Flow** for that specific meal.
-- **State Synchronization**: Verified that selecting a meal and starting the preparation updates the app-wide state reactively.
-
-## Technical Details
-- **Architecture**: Leveraged cross-module dependencies to keep meal cards self-contained in the `:features:health:meals` module while exposing them to the routines feature.
-- **Build Status**: Verified with a successful clean build: `:applications:kocolor:apps:mobile:assembleDebug`.
-
----
-> [!SUCCESS]
-> Your Nutrition Strategy is now truly actionable. Tap a **Meal Suggestion** in your ritual deep-dive to see the details and tap **"Start Cooking"** to begin your preparation protocol.
-
-**KoColor now provides a complete, high-precision journey from biological theory to nutritional reality.**
