+++ features/health/nutrition/src/main/java/com/zoewave/probase/features/health/nutrition/data/MealModels.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.meals.data
+package com.zoewave.probase.features.health.nutrition.data
 
 import kotlinx.serialization.Serializable
