+++ features/health/nutrition/src/main/java/com/zoewave/probase/features/health/nutrition/ui/ritual/NutritionUiState.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.nutrition.ui
+package com.zoewave.probase.features.health.nutrition.ui.ritual
 
 import com.zoewave.probase.core.model.ritual.BeautyRoutine
