+++ features/health/nutrition/src/main/java/com/zoewave/probase/features/health/nutrition/ui/shared/BioNutritionRitualStep.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.nutrition.ui.components
+package com.zoewave.probase.features.health.nutrition.ui.shared
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.border
