+++ features/health/nutrition/src/main/java/com/zoewave/probase/features/health/nutrition/ui/shared/BioOptimizedColors.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.meals.ui
+package com.zoewave.probase.features.health.nutrition.ui.shared
 
 import androidx.compose.ui.graphics.Color
