+++ settings.gradle.kts
@@ -44,7 +44,6 @@ include(":features:health:core")
 include(":features:health:cgm")
 include(":features:health:hydration")
 include(":features:health:nutrition")
-include(":features:health:meals")
 include(":features:camera")
 include(":features:microphone")
 include(":features:calendar")
